LayoutTests/fast/loader/allow-redirect-to-invalid-url-using-meta-refresh-expected.txt

frame "frame" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
frame "frame" - didCommitLoadForFrame
frame "frame" - didFinishDocumentLoadForFrame
frame "frame" - didHandleOnloadEventsForFrame
main frame - didHandleOnloadEventsForFrame
frame "frame" - willPerformClientRedirectToURL: x-dummy://A=a&B=b 
frame "frame" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
Tests that we allow redirection to an invalid URL initiated by <meta http-equiv="refresh"> when setting allowNavigationToInvalidURL is true. This test PASSED if you see an entry in the dumped frame load callbacks of the form: "willPerformClientRedirectToURL: x-dummy://A=a&B=b".

Note, this test must be run in DumpRenderTree.
